refactor(adapters): split blend/kinetic into index/fetch/score/types and docs into folders refactor(adapters): split blend/kinetic tests to mirror the source layout
