+++ .github/ISSUE_TEMPLATE/adapter.yml
@@ -6,7 +6,7 @@ body:
     attributes:
       value: |
         Adapters are the main contribution path. Read
-        [CONTRIBUTING.md](../CONTRIBUTING.md) and [METHODOLOGY.md](../METHODOLOGY.md) first —
+        [CONTRIBUTING.md](../CONTRIBUTING.md) and [methodology/index.md](../methodology/index.md) first —
         thresholds must be anchored to the protocol's own on-chain parameters, never invented.
 
         Any PR that comes out of this should target the `dev` branch, not `main`.
