+++ adapters/index.ts
@@ -1,4 +1,6 @@
-// Protocol adapters live here, one file per protocol (e.g. blend.ts),
-// each implementing the Adapter interface from @stenion/core.
-export * from './blend';
-export * from './kinetic';
+// Protocol adapters live here, one FOLDER per protocol (blend/, kinetic/),
+// each implementing the Adapter interface from @stenion/core. A folder's
+// index.ts is its whole public surface; index/fetch/score/types beside it are
+// internal wiring.
+export * from './blend/index.ts';
+export * from './kinetic/index.ts';
