+++ adapters/tsconfig.build.json
@@ -1,11 +1,20 @@
 {
-  // The emitting config. Excludes tests so they're typechecked but never
-  // compiled into dist/.
+  // The emitting config. Excludes tests — and fixtures, which are test data —
+  // so they're typechecked but never compiled into dist/.
+  //
+  // `allowImportingTsExtensions` + `rewriteRelativeImportExtensions` are here
+  // for the same reason as in indexer/: an adapter's index.ts is BOTH imported
+  // by a test and importing siblings of its own. Node's type-stripping ESM
+  // loader resolves a test's import graph literally, so blend/index.ts must name
+  // `./score.ts`, while the emitted CommonJS must say `./score.js`; the rewrite
+  // flag is what lets one source satisfy both.
   "extends": "../tsconfig.node.json",
   "compilerOptions": {
     "outDir": "dist",
-    "rootDir": "."
+    "rootDir": ".",
+    "allowImportingTsExtensions": true,
+    "rewriteRelativeImportExtensions": true
   },
-  "include": ["*.ts"],
-  "exclude": ["dist", "node_modules", "*.test.ts"]
+  "include": ["**/*.ts"],
+  "exclude": ["dist", "node_modules", "**/*.test.ts", "fixtures"]
 }
