+++ architecture/index.md
@@ -1,0 +1,5 @@
+# Stenion Architecture
+
+The technical shape of the system: how the code is organized, how data flows from the chain to a
+score on the dashboard, and how it's deployed. For _how a score is calculated_, see
+[`methodology/index.md`](../methodology/index.md); for _how to add a protocol_, see [`CONTRIBUTING.md`](../CONTRIBUTING.md).
