+++ dashboard/app/docs/api/page.tsx
@@ -4,6 +4,7 @@ import path from 'node:path';
 import { ExternalLink, FileWarning } from 'lucide-react';
 import { MarkdownDoc } from '../../../components/markdown-doc';
 import { API_DOCS_SOURCE_URL } from '../../lib/site';
+import { API_PARTS } from '../../lib/doc-parts.mjs';
 import { Reveal } from '../../../components/reveal';
 
 export const metadata: Metadata = {
@@ -14,13 +15,27 @@ export const metadata: Metadata = {
 
 // Same shape as /methodology: the doc lives at the repo root as the single
 // source of truth (it's what a developer lands on first from the README, and
-// GitHub renders it), and this route renders that same file rather than keeping
-// a second copy in the app. next.config pins outputFileTracingRoot to the repo
-// root and includes API.md for this route, so it's in the serverless bundle.
+// GitHub renders it), and this route renders those same files rather than
+// keeping a second copy in the app. next.config pins outputFileTracingRoot to
+// the repo root and traces every part for this route, so they're in the bundle.
+//
+// Joined on a single newline, and all-or-nothing on a missing part — see
+// /methodology's loader for both reasons.
 async function loadApiDocs(): Promise<string | null> {
   try {
-    const p = path.join(process.cwd(), '..', 'API.md');
-    return await readFile(p, 'utf8');
+    const parts = await Promise.all(
+      // `turbopackIgnore` because the part name is a variable now, and Turbopack's
+      // static analysis answers "unknown path" by tracing the WHOLE repo into the
+      // serverless output — a build warning, a slower deploy, and a real risk of
+      // the size limit. What actually puts these files in the bundle is the
+      // route's `outputFileTracingIncludes` entry (#96), which names every part
+      // explicitly from this same manifest; the analyser's guess was never what
+      // carried them.
+      API_PARTS.map((part) =>
+        readFile(path.join(/* turbopackIgnore: true */ process.cwd(), '..', part), 'utf8'),
+      ),
+    );
+    return parts.join('\n');
   } catch {
     return null;
   }
@@ -55,7 +70,7 @@ export default async function ApiDocsPage() {
 
       <Reveal delay={0.08} className="mt-12">
         {source ? (
-          <MarkdownDoc source={source} />
+          <MarkdownDoc source={source} basePath="api-docs" />
         ) : (
           <div className="rounded-xl border border-danger/25 bg-danger/5 p-8 text-center">
             <FileWarning className="mx-auto h-8 w-8 text-danger" />
