+++ dashboard/components/markdown-doc.tsx
@@ -24,10 +24,34 @@ function slug(s: string): string {
     .replace(/\s+/g, '-');
 }
 
+/**
+ * Resolve a repo-relative link against the folder the doc it was written in
+ * lives in, into a path from the repo root.
+ *
+ * Needed since the rendered docs became folders: `methodology/lending.md` links
+ * to a sibling as `publishing-rules.md` and to a repo file as
+ * `../core/src/weights.ts`, and both must resolve the same way here as they do
+ * on GitHub. Before the split every doc sat at the repo root, so a link was
+ * already root-relative and this was a no-op — which is what an empty
+ * `basePath` still gives.
+ */
+function resolveRepoPath(href: string, basePath: string): string {
+  const out: string[] = [];
+  for (const segment of [...basePath.split('/'), ...href.split('/')]) {
+    if (segment === '' || segment === '.') continue;
+    if (segment === '..') out.pop();
+    else out.push(segment);
+  }
+  return out.join('/');
+}
+
 // react-markdown gives no heading ids and leaves links repo-relative. We add
 // slug ids (so the doc's own in-page anchors work) and rewrite relative links to
-// the GitHub source (so file references like `adapters/blend.ts` resolve).
-const components: Components = {
+// the GitHub source (so file references like `adapters/blend/index.ts` resolve).
+//
+// A function of `basePath` rather than a module constant, because the `a`
+// override needs it and every other override is independent of it.
+const makeComponents = (basePath: string): Components => ({
   h1: ({ children }) => <h1 id={slug(toText(children))}>{children}</h1>,
   h2: ({ children }) => <h2 id={slug(toText(children))}>{children}</h2>,
   h3: ({ children }) => <h3 id={slug(toText(children))}>{children}</h3>,
@@ -67,28 +91,34 @@ const components: Components = {
       );
     }
     // A repo-relative link. Most resolve to the GitHub source, because most of
-    // the docs a rendered page references (ARCHITECTURE.md, an adapter file) only
+    // the docs a rendered page references (architecture/, an adapter file) only
     // exist there. The exception is a doc that IS rendered on this site: sending
     // a reader from /docs/api out to raw markdown on GitHub for the methodology
     // is worse than keeping them here, so those few map to their own route.
     // Written as one relative link so the same file still resolves on GitHub.
-    const bare = h.replace(/^\.?\//, '');
-    const [file, hash] = bare.split('#');
+    const [rawPath, hash] = h.split('#');
+    const file = resolveRepoPath(rawPath, basePath);
     const route = RENDERED_DOC_ROUTES[file];
     if (route) return <a href={hash ? `${route}#${hash}` : route}>{children}</a>;
     return (
-      <a href={`${REPO_BLOB}/${bare}`} target="_blank" rel="noreferrer">
+      <a href={`${REPO_BLOB}/${file}${hash ? `#${hash}` : ''}`} target="_blank" rel="noreferrer">
         {children}
       </a>
     );
   },
-};
+});
 
-/** Render a markdown string with the Stenion dark prose theme. */
-export function MarkdownDoc({ source }: { source: string }) {
+/**
+ * Render a markdown string with the Stenion dark prose theme.
+ *
+ * `basePath` is the repo folder the source was assembled from ('methodology',
+ * 'api-docs'), and is what the doc's own relative links resolve against. Omit it
+ * for a doc that lives at the repo root.
+ */
+export function MarkdownDoc({ source, basePath = '' }: { source: string; basePath?: string }) {
   return (
     <div className="prose prose-stenion max-w-none prose-headings:font-display prose-headings:tracking-tight prose-h1:text-3xl prose-h2:mt-12 prose-h2:border-t prose-h2:border-line prose-h2:pt-8 prose-a:no-underline hover:prose-a:underline">
-      <Markdown remarkPlugins={[remarkGfm]} components={components}>
+      <Markdown remarkPlugins={[remarkGfm]} components={makeComponents(basePath)}>
         {source}
       </Markdown>
     </div>
